MNT Move timeout handling to subprocess.py, remove sigchld handling altogether.
