+++ lute/execution/executor.py
@@ -32,12 +32,10 @@ import subprocess
 import time
 import os
 import signal
-from typing import Dict, Callable, List, Union, Any, Tuple, Optional
-from typing_extensions import Self, TypeAlias
+from typing import Dict, Callable, List, Optional
+from typing_extensions import Self
 from abc import ABC, abstractmethod
 import warnings
-import types
-import resource
 import copy
 
 from .ipc import *
@@ -441,39 +439,3 @@ class Executor(BaseExecutor):
         reporting to third party services, etc.
         """
         self._task_loop(proc)  # Perform a final read.
-
-
-_SIGNUM: TypeAlias = Union[int, signal.Signals]
-_HANDLER: TypeAlias = Union[
-    Callable[[int, Union[types.FrameType, None]], Any], int, signal.Handlers, None
-]
-
-
-def get_executor(where: str) -> Optional[Executor]:
-    """Return the current Executor."""
-    objects: Dict[str, Any] = globals()
-    for _, obj in objects.items():
-        if isinstance(obj, Executor):
-            return obj
-    return None
-
-
-def sigchld_handler(signum: _SIGNUM, frame: types.FrameType) -> None:
-    """Handle child Task suspension and resumption from outside of Executor."""
-    # (pid, status, resource usage - can maybe infer errors, etc.)
-    ret: Tuple[int, int, resource.struct_rusage] = os.wait4(
-        -1, os.WUNTRACED | os.WCONTINUED
-    )
-    if os.WIFCONTINUED(ret[1]):
-        executor: Optional[Executor] = get_executor(__name__)
-        if executor:
-            executor._analysis_desc.task_result.task_status = TaskStatus.RUNNING
-            logger.info("Task resumed.")
-    elif os.WIFSTOPPED(ret[1]):
-        executor: Optional[Executor] = get_executor(__name__)
-        if executor:
-            executor._analysis_desc.task_result.task_status = TaskStatus.STOPPED
-            logger.info("Task stopped.")
-
-
-# signal.signal(signal.SIGCHLD, sigchld_handler)
